pyproject.toml
[project]
name = "pms-v3"
version = "0.1.0"
description = "PMS V3 Project"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "django-environ>=0.11.2",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "django-celery-beat>=2.5.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
